CB-99 refactor: API slices modularity
rtk query 관련 slice 모듈화
- apiSlice로 공통 config 설정
- 각 endpoint 별 slice를 `@/store/api/`에 구성
