+++ src/store/api/types.ts
@@ -1,0 +1,4 @@
+export interface IGenericResponse {
+  status: string;
+  message: string;
+}
