--- src/utils/api/types.ts
@@ -1,4 +1,0 @@
-export interface IGenericResponse {
-  status: string;
-  message: string;
-}
